refactor: SurveyFormBody -> SurveyForm 이름, 내부 로직 변경 #137
